Don't forward unnecessary headers (unless trying to disable caching)
I misunderstood the behavior that results when you specify headers to
forward: it's not necessary to specify headers in order to have CF
forward them to the origin; CF will do that with many headers
automatically
(https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/RequestAndResponseBehaviorCustomOrigin.html#request-custom-headers-behavior).
CF will cache objects separately when those objects are fetched with
differing header values.

Turns out that fixing the spelling of "Referer" in the prior commit
actually decreased cache hits because CF was now caching different
copies of the same object all because the requests for that object had
different Referer header values.

The 'Host' header stays because I want CF to ask for 'www.' and not
'www0.' (default is to set Host to the FQDN of the origin).