Add "p" query string var to cache whitelist
Fetching a Wordpress URI like "/?p=1128" results in an HTTP 301 redirect
to the permalink URI of post 1128. If "p" isn't whitelisted, then CF
caches the redirect and serves up the redirect for any requests to "/"
which is not the intended behavior and essentially breaks the front page
of the site.

This commit instructs CF to treat "/?p=1128" as a separate object from
"/".