Fix pattern matching for /wordpress/wp-* URIs
Turns out I misunderstood the CF docs when it came to matching the
request URI: it looks like CF matches left-to-right (with an optional
leading '/'). A pattern like 'wp-admin' doesn't match a URI of
'/wordpress/wp-admin'. This was resulting in CF actually applying the
default cache behavior to /wordpress/wp-* URIs (which badly broke the
admin interface after upgrading from 4.9.9 to 5.0.3).

This commit adds '/wordpress/' to the patterns where appropriate.

TODO: be more flexible in the pattern and allow URIs that aren't rooted
in /wordpress.